370: doppelte Eingabe bei Bodenwertaufteilung
Task-Url: http://polymap.org/kaps/ticket/370

Ignoriere Event für RWZ, wenn bereits ausgewähltes RWZ dem neuen RWZ
entspricht, so dass der Bodenrichtwert nicht wieder zurückgesetzt wird

+++ plugins/org.polymap.kaps/src/org/polymap/kaps/ui/MyNumberValidator.java
@@ -122,7 +122,7 @@ public class MyNumberValidator
                 return Float.valueOf( result.floatValue() );
             }
             else if (Double.class.isAssignableFrom( targetClass )) {
-                return Double.valueOf( result.floatValue() );
+                return Double.valueOf( result.doubleValue() );
             }
             else if (Integer.class.isAssignableFrom( targetClass )) {
                 return Integer.valueOf( result.intValue() );
